+++ saleor/graphql/page/tests/queries/test_page.py
@@ -134,6 +134,85 @@ def test_staff_query_unpublished_page_by_id_without_required_permission(
     assert content["data"]["page"] is None
 
 
+def test_app_query_unpublished_page_by_id(
+    app_api_client, page, permission_manage_pages
+):
+    # given
+    page.is_published = False
+    page.save()
+
+    app_api_client.app.permissions.add(permission_manage_pages)
+
+    variables = {"id": graphene.Node.to_global_id("Page", page.id)}
+
+    # when
+    response = app_api_client.post_graphql(
+        PAGE_QUERY,
+        variables,
+    )
+
+    # then
+    content = get_graphql_content(response)
+    assert content["data"]["page"]["id"] == variables["id"]
+
+
+def test_app_query_unpublished_page_by_id_without_required_permission(
+    app_api_client,
+    page,
+):
+    # given
+    page.is_published = False
+    page.save()
+
+    variables = {"id": graphene.Node.to_global_id("Page", page.id)}
+
+    # when
+    response = app_api_client.post_graphql(PAGE_QUERY, variables)
+
+    # then
+    content = get_graphql_content(response)
+    assert content["data"]["page"] is None
+
+
+def test_app_query_unpublished_page_by_slug(
+    app_api_client, page, permission_manage_pages
+):
+    # given
+    page.is_published = False
+    page.save()
+
+    app_api_client.app.permissions.add(permission_manage_pages)
+
+    variables = {"slug": page.slug}
+
+    # when
+    response = app_api_client.post_graphql(
+        PAGE_QUERY,
+        variables,
+    )
+
+    # then
+    content = get_graphql_content(response)
+    assert content["data"]["page"]["id"] == graphene.Node.to_global_id("Page", page.id)
+
+
+def test_app_query_unpublished_page_by_slug_without_required_permission(
+    app_api_client,
+    page,
+):
+    # given
+    page.is_published = False
+    page.save()
+
+    # when
+    variables = {"slug": page.slug}
+
+    # then
+    response = app_api_client.post_graphql(PAGE_QUERY, variables)
+    content = get_graphql_content(response)
+    assert content["data"]["page"] is None
+
+
 def test_staff_query_unpublished_page_by_slug(
     staff_api_client, page, permission_manage_pages
 ):