+++ saleor/graphql/page/tests/queries/test_pages.py
@@ -163,3 +163,132 @@ def test_query_pages_with_sort(
 
     for order, page_name in enumerate(result_order):
         assert pages[order]["node"]["title"] == page_name
+
+
+PAGES_QUERY = """
+    query {
+        pages(first: 10) {
+            edges {
+                node {
+                    id
+                    title
+                    slug
+                    pageType {
+                        id
+                    }
+                    content
+                    contentJson
+                    attributes {
+                        attribute {
+                            slug
+                        }
+                        values {
+                            id
+                            slug
+                        }
+                    }
+                }
+            }
+        }
+    }
+"""
+
+
+def test_query_pages_by_staff(
+    staff_api_client, page_list, page, permission_manage_pages
+):
+    """Ensure staff user with manage pages permission can query all pages,
+    including unpublished pages."""
+    # given
+    unpublished_page = page
+    unpublished_page.is_published = False
+    unpublished_page.save(update_fields=["is_published"])
+
+    page_count = Page.objects.count()
+
+    staff_api_client.user.user_permissions.add(permission_manage_pages)
+
+    # when
+    response = staff_api_client.post_graphql(PAGES_QUERY)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["pages"]["edges"]
+    assert len(data) == page_count
+
+
+def test_query_pages_by_app(app_api_client, page_list, page, permission_manage_pages):
+    """Ensure app with manage pages permission can query all pages,
+    including unpublished pages."""
+    # given
+    unpublished_page = page
+    unpublished_page.is_published = False
+    unpublished_page.save(update_fields=["is_published"])
+
+    page_count = Page.objects.count()
+
+    app_api_client.app.permissions.add(permission_manage_pages)
+
+    # when
+    response = app_api_client.post_graphql(PAGES_QUERY)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["pages"]["edges"]
+    assert len(data) == page_count
+
+
+def test_query_pages_by_staff_no_perm(staff_api_client, page_list, page):
+    """Ensure staff user without manage pages permission can query
+    only published pages."""
+
+    # given
+    unpublished_page = page
+    unpublished_page.is_published = False
+    unpublished_page.save(update_fields=["is_published"])
+
+    page_count = Page.objects.count()
+
+    # when
+    response = staff_api_client.post_graphql(PAGES_QUERY)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["pages"]["edges"]
+    assert len(data) == page_count - 1
+
+
+def test_query_pages_by_app_no_perm(app_api_client, page_list, page):
+    """Ensure app without manage pages permission can query only published pages."""
+    # given
+    unpublished_page = page
+    unpublished_page.is_published = False
+    unpublished_page.save(update_fields=["is_published"])
+
+    page_count = Page.objects.count()
+
+    # when
+    response = app_api_client.post_graphql(PAGES_QUERY)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["pages"]["edges"]
+    assert len(data) == page_count - 1
+
+
+def test_query_pages_by_customer(api_client, page_list, page):
+    """Ensure customer user can query only published pages."""
+    # given
+    unpublished_page = page
+    unpublished_page.is_published = False
+    unpublished_page.save(update_fields=["is_published"])
+
+    page_count = Page.objects.count()
+
+    # when
+    response = api_client.post_graphql(PAGES_QUERY)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["pages"]["edges"]
+    assert len(data) == page_count - 1